Read neuronal skeletons via neuprint

Read neuronal skeletons via neuprint

Usage

read_dyak_neurons(
  ids,
  connectors = FALSE,
  units = c("nm", "raw", "microns"),
  heal.threshold = Inf,
  conn = NULL,
  ...,
  dataset = dyak_default_dataset()
)

Arguments

ids

bodyids in any form compatible with malevnc::manc_ids().

connectors

Whether to fetch synaptic connections for the neuron (default FALSE in contrast to neuprintr::neuprint_read_neurons()).

units

Units of the returned neurons (default nm).

heal.threshold

The threshold for healing disconnected skeleton fragments. The default of Inf ensures that all fragments are joined together.

conn

Optional, a neuprint_connection object. Defaults to dyak_neuprint() to ensure that query is against the yakuba dataset.

...

Additional arguments passed to neuprintr::neuprint_read_neurons().

dataset

The name of the dataset as reported in Clio or neuprint.

Value

A nat::neuronlist object containing one or more neurons.

See also

Examples

# \donttest{
dns <- read_dyak_neurons("DNa02")
#> Warning: NAs introduced by coercion to integer64 range
dns
#> 'neuronlist' containing 2 'neuprintneuron' objects and 'data.frame' with 39 vars [158.7 kB]
# }