yakuba: Access the Drosophila yakuba FlyEM Dataset

Provides a thin wrapper around the 'malevnc' package for working with the Drosophila yakuba male VNC neuprint and annotation datasets.

Bridging registrations

yakuba registers a bundled thin-plate spline transform from yakubaum to MANC at package startup when nat.templatebrains is available. A simple scaling alias is also added from yakuba (nm) to yakubaum (microns), so nat.templatebrains::xform_brain() can chain from yakuba to MANC.

Note the current registration (which lives in inst/extdata) was computed by Sebastian Cachero in April 2025. See slack for details

Package Options

There is currently one package option:

yakuba.dataset keeps track of the active yakuba dataset.

You can use this to run your code against the production yakuba dataset or future snapshot variants. The main ways to do this are:

Use with_dyak() to run one expression without changing the default yakuba dataset.

Use choose_dyak_dataset() to choose a default yakuba dataset for the rest of the session.

Expert users may set the yakuba.dataset option directly in their .Rprofile.

See also

Author

Maintainer: Gregory Jefferis jefferis@gmail.com (ORCID)

Examples

# \donttest{
options()[grepl("^yakuba", names(options()))]
#> $yakuba.dataset
#> [1] "yakuba"
#> 
# }

# \donttest{
library(malevnc)
library(nat)
#> Loading required package: rgl
#> Some nat functions depend on a CMTK installation. See ?cmtk and README.md for details.
#> 
#> Attaching package: ‘nat’
#> The following object is masked from ‘package:rgl’:
#> 
#>     wire3d
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, union
library(nat.templatebrains)

MANC.tissue.surf.yak = nat.templatebrains::xform_brain(
  MANC.tissue.surf,
  sample = "MANC",
  reference = "yakuba"
)
wire3d(MANC.tissue.surf.yak)

3D plot

dna02 = read_dyak_neurons("DNa02")
#> Warning: NAs introduced by coercion to integer64 range
plot3d(dna02, lwd = 3)

# nb the default template space for MANC is calibrated in microns
# but for yakuba we use nm
dna02.manc = malevnc::manc_read_neurons("DNa02", units = "microns")
#> Warning: NAs introduced by coercion to integer64 range
dna02.manc.yak = nat.templatebrains::xform_brain(
  dna02.manc,
  reference = "yakuba",
  sample = "MANC"
)
# set up nice coords for the plot
plot3d(dna02.manc.yak, col = "black", lwd = 2)
#' # }